Many Analytics changes
