+++ src/utils/analytics/types.ts
@@ -1,0 +1,7 @@
+
+export interface AnalyticsEvent {
+  category: string
+  label: string
+  action: string
+  value?: string | number
+}
