+++ src/utils/analyticsEvents.ts
@@ -1,4 +1,4 @@
-import { AnalyticsEvent } from "./AnalyticsManager"
+import { AnalyticsEvent } from "./analytics/types"
 
 export const staticEvents = {
   START_AUTOMATIC_CHALLENGER: {
